Ticket:22 In removing the directory information from the parity objects; the entire xml io layer had to be re-created.

+++ client-model/src/test/java/com/thinkparity/model/ModelTestCaseHelper.java
@@ -76,7 +76,8 @@ public class ModelTestCaseHelper {
 		// set the workspace directory, then delete its contents
 		final File jUnitWorkspace =
 			new File(resourcesDirectory, "junit.workspace");
-		FileUtil.deleteTree(jUnitWorkspace);
+		if(jUnitWorkspace.exists())
+			FileUtil.deleteTree(jUnitWorkspace);
 		// set the test files
 		final File jUnitResourcesFiles =
 			new File(resourcesDirectory, "junit.files");
